requirements.txt
django
djangorestframework
pymysql
mysqlclient
